requirements.txt
# Base dependencies. Modules might have additional requirements.
# It is recommended to pin versions for all packages (e.g., using 'pip freeze').
openwakeword
openai-whisper
pyaudio
pyttsx3
pvporcupine
numpy
speedtest-cli
requests
psutil
pytest
python-dotenv